\name{DEoptimLoop} 
\alias{DEoptimLoop} 
\title{likeLTD::DEoptimLoop} 
\usage{DEoptimLoop(PARAMS, tolerance=1e-6)}
\arguments{
  \item{PARAMS}{Parameters for optimization as generated by optimization.params}
  \item{tolerance}{If the relative difference between the current result, and the last checked result is less than this value, then it is classed as converged.}
}
\value{List.} 
\description{Values returned by DEoptim after optimization. See DEoptim help page for more details.}
\details{
  Optimize over parameter space, while checking for converence every 50 iterations.
}
\seealso{DEoptim}
