\name{DNA17-db}
\alias{DNA17-db}
\docType{data}
\title{NGMSelect allele database }
\description{A table with allele frequencies for NGMSelect loci. 
Populations are "NDU1" (Caucasian), "NDU2" (African + Afro-Caribbean), 
"NDU3" (South Asian), "NDU4" (East Asian), "NDU6" (African), 
"NDU7" (Afro-Caribbean).}
\format{A table, as described in \code{load.allele.database}.}
\keyword{dataset}
