\name{DNA17-lus-db}
\alias{DNA17-lus-db}
\docType{data}
\title{NGMSelect allele database with LUS values}
\description{A table with allele frequencies for NGMSelect loci, 
including the longest uninterrupted sequence (LUS).}
\format{A table, as described in \code{load.allele.database}, with LUS.}
\keyword{dataset}
