\name{evaluate.from.interim} 
\alias{evaluate.from.interim} 
\title{likeLTD::evaluate.from.interim} 
\usage{evaluate(file)}
\arguments{
  \item{file}{restricted to only accept a filepath including the name "interim.RData", which is generated by evaluate() only if interim=TRUE. "interim.RData" will also be subsequently generated by evaluate.from.interim() thereafter, at regular intervals during the evaluation}
}
\value{
A list containing three elements:
\item{Pros}{Prosecution results, structured as results from DEoptim::DEoptim.}
\item{Def}{Defence results, structured as results from DEoptim::DEoptim.}
\item{WoE}{WoE for each chunk. The final value if the final WoE.}
} 
\description{
Optimize both prosecution and defence likelihoods, returning the weight of evidence.
}
\details{
  Optimize over parameter space, using a geometric progression of crossover rate and tolerance.
Both prosecution and defence cases are optimized simultaneously.
}
\seealso{evaluate}
