\name{evaluate} 
\alias{evaluate} 
\title{likeLTD::evaluate} 
\usage{evaluate(P.pars, D.pars, tolerance=1e-5, n.chunks=NULL, progBar=TRUE)}
\arguments{
  \item{P.pars}{Parameters for prosecution hypothesis, as generated by optimisation.params.}
\item{D.pars}{Parameters for prosecution hypothesis, as generated by optimisation.params.}
\item{tolerance}{Tolerance for the final chunk of optimisation. If the relative difference 
between the current result, and the last checked result is less than this value, then it is 
classed as converged.}
\item{n.chunks}{Number of chunks to run. Defaults to NULL. if n.chunks is NULL , the number
 of chunks to run is determined by the mean of the standard deviation of the initial phase of 
the intial chunk of optimisation for prosecution and defence. }
\item{progBar}{Logical, stating whether to display a graphical progress basr or not. This should
be set to FALSE is the user does not have graphical capabilities e.g. if running from command
line on a server.}

}
\value{List.} 
\description{
\item{Pros}{Prosecution results, structured as results from DEoptim::DEoptim.}
\item{Def}{Defence results, structured as results from DEoptim::DEoptim.}
\item{WoE}{WoE for each chunk. The final value if the final WoE.}
}
\details{
  Optimize over parameter space, using a geometric progression of crossover rate and tolerance.
Both prosecution and defence cases are optimized simultaneously.
}
\seealso{DEoptim,DEoptimLoop}
