\name{Identifiler-db}
\alias{Identifiler-db}
\docType{data}
\title{Identifiler allele database }
\description{A table with allele frequencies for Identifiler loci.
Populations are "EA1" (Caucasian), "EA3" (African/Afro-Caribbean), 
"EA4" (South Asian).}
\format{A table, as described in \code{load.allele.database}.}
\keyword{dataset}
