\name{likeLTD}
\alias{likeLTD}
\title{likeLTD}
\docType{package}
\description{A package to compute likelihoods for low template DNA profiles.}
\details{
  \pkg{likeLTD} ("likelihoods for Low Template DNA proﬁles") is an R package
  for computing likelihoods for DNA proﬁles. It is freely available under a GNU
    public license.

  It is particularly suited for low-template and/or degraded DNA when alleles from
  some contributors may be subject to dropout. It can handle any number of proﬁled
  possible contributors and up to three unproﬁled contributors.
}
\references{David J. Balding, John Buckleton, "Interpreting low template DNA
profiles", \emph{Forensic Science International: Genetics}, \bold{4}, 1--10
(2009)}
