\name{linkage}
\alias{linkage}
\docType{data}
\title{Recombination rates for linked loci}
\description{A table recombination rates for a set of linked loci.}
\format{An nxn table, where n is the number of loci in the database.
 If locus x and y are linked, then table[x,y] and table[y,x] contain 
 the recombination rate for those loci. Unlinked loci have value NA.}
\keyword{dataset}
