\name{locus.likes.peaks} 
\alias{locus.likes.peaks} 
\title{likeLTD::locus.likes.peaks} 
\usage{locus.likes.peaks(hypothesis,results,\dots)}
\arguments{
  \item{hypothesis}{The hypothesis generated by either \code{prosecution.hypothesis.peaks} 
  or \code{defence.hypothesis.peaks}.}
  \item{results}{Either prosecution or defence results from \code{evaluate.peaks}
  e.g. \code{results$Pros} or \code{results$Def}.}
  \item{\dots}{Any extra parameter that was passed to \code{evaluate.peaks}.}
}
\value{Vector.} 
\description{Vector with individual locus likelihoods for peak height data.}
\details{
  Convert the overall likelihood returned by \code{evaluate.peaks} into locus specific likelihoods.
}
\seealso{evaluate.peaks}
