\name{locus.likes} 
\alias{locus.likes} 
\title{likeLTD::locus.likes} 
\usage{locus.likes(hypothesis,results,\dots)}
\arguments{
  \item{hypothesis}{The hypothesis generated by either \code{prosecution.hypothesis} 
  or \code{defence.hypothesis}.}
  \item{results}{Either prosecution or defence results from \code{evaluate} 
e.g. \code{results$Pros} or \code{results$Def}.}
  \item{\dots}{Any extra parameter that was passed to \code{evaluate} or 
\code{optimisation.params}. }
}
\value{Vector.} 
\description{Vector with individual locus likelihoods.}
\details{
  Convert the overall likelihood returned by DEoptim into locus specific likelihoods.
}
\seealso{evaluate}
