\name{locusLikes} 
\alias{locusLikes} 
\title{likeLTD::locusLikes} 
\usage{locusLikes(hypothesis,results)}
\arguments{
  \item{hypothesis}{The hypothesis generated by either prosecution.hypothesis() or defense.hypothesis().}
  \item{results}{Results from do.call(optim, parameters), or multiRun(hypothesis, nrun, ...)$bestResults.}
}
\value{Vector.} 
\description{Vector with individual locus likelihoods.}
\details{
  Convert the overall likelihood returned by optim into locus specific likelihoods.
}
\seealso{multiRun}
