\name{NISTglobalfiler-db}
\alias{NISTglobalfiler-db}
\docType{data}
\title{Globalfiler allele database from NIST}
\description{Globalfiler allele database from the 
National Institute of Standards and Technology (NIST). 
See http://strbase.nist.gov/NISTpop.htm.

A table with allele frequencies for globalfiler loci. 
Populations are "AfAm" (African-American), "Cau" (Caucasian),"Hisp" (Hispanic), "Asian".}
\format{A table, as described in \code{load.allele.database}.}
\keyword{dataset}
