\name{NISTidentifiler-db}
\alias{NISTidentifiler-db}
\docType{data}
\title{Identifiler allele database from NIST}
\description{Identifiler allele database from the 
National Institute of Standards and Technology (NIST). 
See http://strbase.nist.gov/NISTpop.htm.

A table with allele frequencies for identifiler loci. 
Populations are "AfAm" (African-American), "Cau" (Caucasian),"Hisp" (Hispanic)}
\format{A table, as described in \code{load.allele.database}.}
\keyword{dataset}
