\name{SGMplus-db}
\alias{SGMplus-db}
\docType{data}
\title{SGMplus allele database }
\description{A table with allele frequencies for SGMplus loci. 
Populations are "EA1" (Caucasian), "EA3" (African/Afro-Caribbean), 
"EA4" (South Asian).}
\format{A table, as described in \code{load.allele.database}.}
\keyword{dataset}
