\name{likeLTD}
\alias{likeLTD}
\title{likeLTD}
\docType{package}
\description{A package to compute likelihoods for low template DNA profiles.}
\details{
  \pkg{likeLTD} (\dQuote{likelihoods for Low Template DNA profiles}) is an R package
  for computing likelihoods for DNA profiles. It is freely available under a
  GNU public license.

  It is particularly suited for low-template and/or degraded DNA when alleles
  from some contributors may be subject to dropout. It can handle any number of 
  profiled possible contributors and up to three unprofiled contributors.
}
\references{David J. Balding, John Buckleton, \dQuote{Interpreting low template DNA
profiles}, \emph{Forensic Science International: Genetics}, \bold{4}, 1--10
(2009)}
