\name{likeLTD}
\alias{likeLTD}
\title{likeLTD}
\docType{package}
\description{A package to compute likelihoods for low template DNA profiles.}
\details{
  \pkg{likeLTD} (\dQuote{likelihoods for Low Template DNA profiles}) is an R package
  for computing likelihoods for DNA profiles. It is freely available under a
  GNU public license.

  It is particularly suited for low-template and/or degraded DNA when alleles
  from some contributors may be subject to dropout. It can handle any number of 
  profiled possible contributors and up to three unprofiled contributors.
  
  \pkg{likeLTD} is currently in beta, please report any issues to the maintainer.
  
  See https://sites.google.com/site/baldingstatisticalgenetics/ for an in-depth guide
  to \pkg{likeLTD}.
}
\references{David J. Balding, \dQuote{Evaluation of mixed-source, low-template dna 
profiles in forensic science}, \emph{Proceedings of the National Academy of Sciences 
of USA}, (2009)}
