\name{locus.likes} 
\alias{locus.likes} 
\title{likeLTD::locus.likes} 
\usage{locus.likes(hypothesis,results,\dots)}
\arguments{
  \item{hypothesis}{The hypothesis generated by either prosecution.hypothesis() or defense.hypothesis().}
  \item{results}{Results from do.call(optim, parameters), or multiRun(hypothesis, nrun, \dots).}
}
\value{Vector.} 
\description{Vector with individual locus likelihoods.}
\details{
  Convert the overall likelihood returned by optim into locus specific likelihoods.
}
\seealso{multiRun}
