\name{locus.likes} 
\alias{locus.likes} 
\title{likeLTD::locus.likes} 
\usage{locus.likes(hypothesis,results,\dots)}
\arguments{
  \item{hypothesis}{The hypothesis generated by either prosecution.hypothesis or defense.hypothesis.}
  \item{results}{Results from optim or multiRun.}
  \item{\dots}{Any extra parameter that was passed to multiRun or optimisation.params e.g. fixed}
}
\value{Vector.} 
\description{Vector with individual locus likelihoods.}
\details{
  Convert the overall likelihood returned by optim into locus specific likelihoods.
}
\seealso{multiRun}
